add utils module & move common types to it
