+++ src/lib.rs
@@ -1,5 +1,5 @@
 use crate::conv::map_instance_descriptor;
-use std::{borrow::Cow, collections::HashMap, ffi::CString, sync::Arc, sync::Mutex};
+use std::{collections::HashMap, ffi::CString, sync::Arc, sync::Mutex};
 use wgc::id;
 
 pub mod command;
@@ -7,6 +7,7 @@ pub mod conv;
 pub mod device;
 pub mod logging;
 pub mod unimplemented;
+pub mod utils;
 
 pub type Context = wgc::hub::Global<wgc::hub::IdentityManagerFactory>;
 
@@ -87,182 +88,6 @@ pub mod native {
     }
 
     include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
-}
-
-pub type Label<'a> = Option<Cow<'a, str>>;
-
-struct OwnedLabel(Option<String>);
-impl OwnedLabel {
-    fn new(ptr: *const std::os::raw::c_char) -> Self {
-        Self(if ptr.is_null() {
-            None
-        } else {
-            Some(
-                unsafe { std::ffi::CStr::from_ptr(ptr) }
-                    .to_string_lossy()
-                    .to_string(),
-            )
-        })
-    }
-    fn into_inner(self) -> Option<String> {
-        self.0
-    }
-    fn as_cow(&self) -> Option<Cow<str>> {
-        self.0.as_ref().map(|s| Cow::Borrowed(s.as_str()))
-    }
-    fn into_cow<'a>(self) -> Option<Cow<'a, str>> {
-        self.0.map(Cow::Owned)
-    }
-}
-
-pub unsafe fn make_slice<'a, T: 'a>(pointer: *const T, count: usize) -> &'a [T] {
-    if count == 0 {
-        &[]
-    } else {
-        std::slice::from_raw_parts(pointer, count)
-    }
-}
-
-/// Follow a chain of next pointers and automatically resolve them to the underlying structs.
-///
-/// # Syntax:
-///
-/// Given:
-///
-/// `fn map_thing_descriptor(base: &ThingDescriptor, ext1: Option<&ThingDescriptorExtension1>) -> wgt::ThingDescriptor`
-///
-/// Use the syntax:
-///
-/// `follow_chain!(map_thing_descriptor(base_c_descriptor, ThingDescriptorExtension1STypeValue => ThingDescriptorExtension1))`
-///
-/// # Safety
-///
-/// This macro does not use any internal unsafe blocks. The caller (or most likely the function) needs
-/// to be unsafe. The following constraints must be upheld for this to be valid:
-///
-/// - All pointers in the chain of next pointers must point to either null or a valid extension object
-/// - All structures used as extension objects must be `#[repr(C)]`.
-/// - All structures used as extension objects must have `pub next_in_chain: Option<&ChainedStruct>` and `pub s_type: SType`
-///   as the first and second members respectively.
-///
-/// The result of these rules, and the fact that wgpu-native functions using it do not validate all these assumptions,
-/// using this macro is an indication that the function itself must be made unsafe.
-///
-/// # Notes
-///
-/// Given two or more extension structs of the same SType in the same chain, this macro will favor the latter most. There should
-/// not be more than one extension struct with the same SType in a chain anyway, so this behavior should be unproblematic.
-
-#[macro_export]
-macro_rules! follow_chain {
-    ($func:ident($base:expr $(, $stype:ident => $ty:ty)*)) => {{
-    #[allow(non_snake_case)] // We use the type name as an easily usable temporary name
-    {
-        $(
-            let mut $stype: Option<&$ty> = None;
-        )*
-        let mut chain_opt: Option<&$crate::native::WGPUChainedStruct> = $base.nextInChain.as_ref();
-        while let Some(next_in_chain) = chain_opt {
-            match next_in_chain.sType {
-                $(
-                    $crate::native::$stype => {
-                        let next_in_chain_ptr = next_in_chain as *const $crate::native::WGPUChainedStruct;
-                        assert_eq!(
-                            0,
-                            next_in_chain_ptr.align_offset(::std::mem::align_of::<$ty>()),
-                            concat!("Chain structure pointer is not aligned correctly to dereference as ", stringify!($ty), ". Correct alignment: {}"),
-                            ::std::mem::align_of::<$ty>()
-                        );
-                        let type_ptr: *const $ty = next_in_chain_ptr as _;
-                        $stype = Some(&*type_ptr);
-                    }
-                )*
-                _ => {}
-            }
-            chain_opt = next_in_chain.next.as_ref();
-        }
-        $func($base, $($stype),*)
-    }}};
-}
-
-/// Creates a function which maps native constants to wgpu enums.
-/// If an error message is provided, the function will panic if the
-/// input does not match any known variants. Otherwise a Result<T, i32> is returned
-///
-/// # Syntax
-///
-/// For enums that have undefined variants:
-/// ```ignore
-/// map_enum!(function_name, header_prefix, rust_type, Variant1, Variant2...)
-/// ```
-///
-/// For enums where all variants are defined:
-/// ```ignore
-/// map_enum!(function_name, header_prefix, rust_type, err_msg, Variant1, Variant2...)
-/// ```
-///
-/// # Example
-///
-/// For the following enum:
-/// ```c
-/// typedef enum WGPUIndexFormat {
-///     WGPUIndexFormat_Undefined = 0x00000000,
-///     WGPUIndexFormat_Uint16 = 0x00000001,
-///     WGPUIndexFormat_Uint32 = 0x00000002,
-///     WGPUIndexFormat_Force32 = 0x7FFFFFFF
-/// } WGPUIndexFormat;
-/// ```
-/// Then you can use the following macro:
-/// ```ignore
-/// map_enum!(map_index_format, WGPUIndexFormat, wgt::IndexFormat, Uint16, Uint32);
-/// ```
-/// Which expands into:
-/// ```ignore
-/// pub fn map_index_format(value: i32) -> Result<wgt::IndexFormat, i32> {
-///      match value {
-///          native::WGPUIndexFormat_Uint16 => Ok(wgt::IndexFormat::Uint16),
-///          native::WGPUIndexFormat_Uint32 => Ok(wgt::IndexFormat::Uint32),
-///          x => Err(x),
-///      }
-/// }
-/// ```
-///
-#[macro_export]
-macro_rules! map_enum {
-    ($name:ident, $c_name:ident, $rs_type:ty, $($variant:ident),+) => {
-        #[inline]
-        pub fn $name(value: native::$c_name) -> Result<$rs_type, native::$c_name> {
-            match value {
-                $(paste::paste!(native::[<$c_name _ $variant>]) => Ok(<$rs_type>::$variant)),+,
-                x => Err(x),
-            }
-        }
-    };
-    ($name:ident, $c_name:ident, $rs_type:ty, $err_msg:literal, $($variant:ident),+) => {
-        #[inline]
-        pub fn $name(value: native::$c_name) -> $rs_type {
-            map_enum!(map_fn, $c_name, $rs_type, $($variant),+);
-
-            map_fn(value).expect($err_msg)
-        }
-    };
-    ($name:ident, $c_name:ident, $rs_type:ty, $($native_variant:ident:$variant2:ident),+) => {
-        #[inline]
-        pub fn $name(value: native::$c_name) -> Result<$rs_type, native::$c_name> {
-            match value {
-                $(paste::paste!(native::[<$c_name _ $native_variant>]) => Ok(<$rs_type>::$variant2)),+,
-                x => Err(x),
-            }
-        }
-    };
-    ($name:ident, $c_name:ident, $rs_type:ty, $err_msg:literal, $($native_variant:ident:$variant2:ident),+) => {
-        #[inline]
-        pub fn $name(value: native::$c_name) -> $rs_type {
-            map_enum!(map_fn, $c_name, $rs_type, $($native_variant:$variant2),+);
-
-            map_fn(value).expect($err_msg)
-        }
-    };
 }
 
 #[no_mangle]
